+++ src/s_stats.c
@@ -19,7 +19,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307
  *  USA
  *
- *  $Id: s_stats.c,v 7.30 2005/08/15 11:45:29 michael Exp $
+ *  $Id: s_stats.c,v 7.31 2005/09/06 15:51:58 adx Exp $
  */
 
 #include "stdinc.h"
@@ -109,12 +109,10 @@ tstats(struct Client *source_p)
              (unsigned int)sp->is_sv);
   sendto_one(source_p, ":%s %d %s T :bytes sent %llu %llu",
              me.name, RPL_STATSDEBUG, source_p->name,
-             (unsigned long long)sp->is_cbs,
-             (unsigned long long)sp->is_sbs);
+             sp->is_cbs, sp->is_sbs);
   sendto_one(source_p, ":%s %d %s T :bytes recv %llu %llu",
              me.name, RPL_STATSDEBUG, source_p->name,
-             (unsigned long long)sp->is_cbr,
-             (unsigned long long)sp->is_sbr);
+             sp->is_cbr, sp->is_sbr);
   sendto_one(source_p, ":%s %d %s T :time connected %u %u",
              me.name, RPL_STATSDEBUG, source_p->name,
              (unsigned int)sp->is_cti,
